requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
mysqlclient==2.2.0
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
django-filter==23.3
django-extensions==3.2.3
python-dateutil==2.8.2
faker==20.1.0
setuptools
pandas